+++ magda-web-client/src/Pages/HomePage.js
@@ -7,15 +7,20 @@ import { config } from "../config.js";
 import "./HomePage.css";
 
 import TagLine from "./HomePageComponents/TagLine";
-import Lozenge from "./HomePageComponents/TagLine";
+import Lozenge from "./HomePageComponents/Lozenge";
 import Stories from "./HomePageComponents/Stories";
+import {Small, Medium} from '../UI/Responsive';
+
+import MediaQuery from "react-responsive";
 
 const getBgImg=()=>{
+    let imageMap = {};
     if(!config.homePageConfig || !config.homePageConfig.backgroundImageUrls) return null;
     const backgroundImageUrls = config.homePageConfig.backgroundImageUrls;
     if(!backgroundImageUrls.length) return null;
     const baseUrl = config.homePageConfig.baseUrl ? config.homePageConfig.baseUrl : "";
-    const imgInfoItems = backgroundImageUrls.map(item=>{
+
+    backgroundImageUrls.forEach(item=>{
         let width;
         try{
             width = parseInt(item.replace(/[^\d]/g,""),10);
@@ -23,26 +28,47 @@ const getBgImg=()=>{
         }catch(e){
             width=0;
         }
-        return {
-            url : baseUrl+item,
-            width
-        }
+
+        imageMap = Object.assign(imageMap, {[width]: baseUrl+item});
     });
-    const srcset = imgInfoItems.filter(item=>item.width).map(item=>`${item.url} ${item.width}w`).join(", ");
-    let sizes;
-    if(imgInfoItems.length<=1) sizes = `${imgInfoItems[0].width}px`;
-    else sizes = imgInfoItems.slice(1).filter(item=>item.width).map(item=>`(min-width: ${item.width}px) ${item.width}px`).join(", ")+`, ${imgInfoItems[0].width}px`;
-    return <img alt="background" className="homepage-background-img" src={imgInfoItems[0].url} sizes={sizes} srcSet={srcset} />;
+
+    const screenSizes = Object.keys(imageMap);
+
+    function getBackgroundImage(imageUrl){
+      return {
+              backgroundImage: 'url(' + imageUrl + ')',
+              backgroundPosition: 'center',
+              backgroundRepeat: 'no-repeat',
+              backgroundSize: 'cover',
+             }
+    }
+    return (<div>{screenSizes.map((size, i) => <MediaQuery
+                  key={size}
+                  minWidth={size + 'px'}
+                  maxWidth={i === screenSizes.length - 1 ? null : screenSizes[i+1] + 'px'}>
+                  <div className='homepage-background-img'
+                       style={getBackgroundImage(imageMap[size])}/>
+                 </MediaQuery>)}</div>);
 };
 
+const getTagLine = ()=>{
+  const homePageConfig = config.homePageConfig;
+  return {
+            desktop: config && homePageConfig.tagLineTextDesktop ? homePageConfig.tagLineTextDesktop : '',
+            mobile: config && homePageConfig.tagLineTextMobile ? homePageConfig.tagLineTextMobile : ''
+          }
+}
+
 const HomePage = withRouter(({ location }) => {
+  
     return (
         <div className="homepage-app-container">
             {getBgImg()}
             <Container className="app-container">
                 <Header/>
+                <Small><TagLine taglineText = {getTagLine().mobile}/></Small>
                 <SearchBox />
-                <TagLine />
+                <Medium><TagLine taglineText = {getTagLine().desktop}/></Medium>
                 <Lozenge />
                 <Stories />
             </Container>